Fix wrong direction on open for some blind types 	modified:   History.txt 	modified:   dzVents/documentation/README.md 	modified:   dzVents/documentation/README.wiki 	modified:   dzVents/documentation/history.md 	modified:   dzVents/runtime/Domoticz.lua 	modified:   dzVents/runtime/Utils.lua 	modified:   dzVents/runtime/device-adapters/generic_device.lua 	modified:   dzVents/runtime/device-adapters/switch_device.lua 	modified:   dzVents/runtime/misc/testdzVents.sh 	modified:   dzVents/runtime/tests/testDevice.lua 	modified:   dzVents/runtime/tests/testUtils.lua 	modified:   main/dzVents.cpp

+++ dzVents/runtime/misc/testdzVents.sh
@@ -88,9 +88,11 @@ checkStarted()
 
 function cleanup
 	{
-		ps -ef | grep domoticz
-		rm domoticz.log[0-9][0-9]*
-		find . -type f -name 'domoticz.db_*' -mmin +30 -exec rm {} \;
+		ps -aux | grep [f]domoticz
+		if [[ $? -eq 1 ]]; then
+			rm domoticz.log[0-9][0-9]*
+			find . -type f -name 'domoticz.db_*' -mmin +30 -exec rm {} \;
+		fi
 	}
 
 function stopBackgroundProcesses
@@ -168,6 +170,9 @@ checkStarted "server" 10
 
 # Just to be sure we do not destroy something important without a backup
 cp $basedir/domoticz.db $basedir/domoticz.db_$$
+
+# Make sure we start with a clean sheet
+rm -f $basedir/domoticz.db
 
 cd $basedir
 ./domoticz > domoticz.log$$ &
